Move from api.model.request to requests
